fixes some error
